Replaced os module with platform

+++ app.py
@@ -1,22 +1,37 @@
 from flask import Flask, request, jsonify
 import psutil
 import json
-import os
 from subprocess import Popen
+import platform
 
-ACTIONS = {
-    "shutdown": ['shutdown', '-s', '-f', '-t', '0'],
-    "hibernate": ['shutdown', '-h'],
-    "logout": ['logoff'],
-    "reboot": ['shutdown', '-r', '-f', '-t', '0'],
-    "lock": ['rundll32', 'user32.dll,LockWorkStation'],
-    "off": ["powershell", "-Command", '(Add-Type -MemberDefinition "[DllImport(""user32.dll"")]`npublic static extern int SendMessage(int hWnd, int hMsg, int wParam, int lParam);" -Name "Win32SendMessage" -Namespace Win32Functions -PassThru)::SendMessage(0xffff, 0x0112, 0xF170, 2)']
-}
 # constants
 HOST = "0.0.0.0"
 PORT = 80
 DEBUG = True
 THREADED = True
+
+#logic to check if OS is Windows
+IS_WINDOWS = platform.system() == "Windows"
+
+
+def actionsBasedOnOS():
+    if IS_WINDOWS:
+        return {
+            "shutdown": ['shutdown', '-s', '-f', '-t', '0'],
+            "hibernate": ['shutdown', '-h'],
+            "logout": ['logoff'],
+            "reboot": ['shutdown', '-r', '-f', '-t', '0'],
+            "lock": ['rundll32', 'user32.dll,LockWorkStation'],
+            "off": ["powershell", "-Command", '(Add-Type -MemberDefinition "[DllImport(""user32.dll"")]`npublic static extern int SendMessage(int hWnd, int hMsg, int wParam, int lParam);" -Name "Win32SendMessage" -Namespace Win32Functions -PassThru)::SendMessage(0xffff, 0x0112, 0xF170, 2)']
+        }
+
+    # Add support for Linux later
+    if platform.system() == "Linux":
+        return {}
+    return {}
+
+
+ACTIONS = actionsBasedOnOS()
 
 
 def fetchSystemDetails():
@@ -48,14 +63,11 @@ def fetchSystemDetails():
     _cpu["cpuUsageTotal%"] = round(
         (sum(cpuUsage)/(100*_cpu["cpuCores"]))*100, 2)
     _cpu["cpuUsage%"] = cpuUsage
-    # cpusUsage as text
-    # _cpu["cpuUsage"] = ", ".join(f"{cu}%"for cu in psutil.cpu_percent(1, True)).strip()
-
     systemDetails.append(_cpu)
 
     _diskUsage = []
     for disk in psutil.disk_partitions():
-        if os.name == 'nt':
+        if IS_WINDOWS:
             if 'cdrom' in disk.opts or disk.fstype == '':
                 continue
         diskDetails = psutil.disk_usage(disk.mountpoint)
@@ -77,8 +89,12 @@ def mapActionToCommand(action):
 
 
 def execCmd(endpoint):
-    Popen(mapActionToCommand(endpoint.lower()))
-    return f'''{endpoint.upper()} command is executed successfully'''
+    cmd = mapActionToCommand(endpoint.lower())
+    if cmd and len(cmd):
+        Popen(cmd)
+        return f'''{endpoint} command is executed successfully'''
+    else:
+        return f'''Failed to map "{endpoint}"" to a command'''
 
 
 if __name__ == "__main__":
@@ -107,15 +123,16 @@ if __name__ == "__main__":
     @app.route('/Lock')
     @app.route('/Off')
     def actions():
-        execCmd((request.path)[1:])
+        command = request.path[1:]
         return '''
         <script>
         window.onbeforeunload = e => {
             e.preventDefault();
             e.returnValue = '';
-        }
-        </script>
-        ''' + f"""{(request.path)[1:].upper()} executed successfully."""
+        };''' + f"""
+        alert('{execCmd(command)}');
+        window.history.back();
+        </script>"""
 
     @app.route('/remote')
     def remote():
@@ -125,13 +142,13 @@ if __name__ == "__main__":
         <title>SysMon</title>
         <meta http-equiv="refresh" content="30">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        |<a href="/Off" target="_blank"><b>&#x260D; </b>TOS</a>
-        |<a href="/Lock" target="_blank"><b>&#x262F; </b>Lock</a>
-        |<a href="/Hibernate" target="_blank"><b>&#x273E; </b>Hibernate</a>
+        |<a href="/Off"><b>&#x260D; </b>TOS</a>
+        |<a href="/Lock"><b>&#x262F; </b>Lock</a>
+        |<a href="/Hibernate"><b>&#x273E; </b>Hibernate</a>
         <br>
-        |<a href="/Logout" target="_blank"><b>&#x26E2; </b>Logout</a>
-        |<a href="/Shutdown" target="_blank"><b>&#x2620; </b>Shutdown</a>
-        |<a href="/Reboot" target="_blank"><b>&#x267A; </b>Reboot</a>
+        |<a href="/Logout"><b>&#x26E2; </b>Logout</a>
+        |<a href="/Shutdown"><b>&#x2620; </b>Shutdown</a>
+        |<a href="/Reboot"><b>&#x267A; </b>Reboot</a>
         <br>
         <br>
         <pre>{details}</pre>
